+++ src/fkooman/Json/JsonException.php
@@ -20,7 +20,6 @@ namespace fkooman\Json;
 
 class JsonException extends \Exception
 {
-
     public function __construct($message, $code = 0, Exception $previous = null)
     {
         switch ($message) {