+++ CHANGES.md
@@ -1,4 +1,10 @@
 # Changelog
 
+## 0.2.0
+- new API
+- no longer use static methods
+- prefer `encode()` and `decode()` now, but aliases `enc()` and `dec()` still 
+  exist
+
 ## 0.1.0
 - initial release